MCP-Server-Agent-Both_Local/requirements.txt
fastmcp>=0.1.0
mistralai>=2.0.0
python-dotenv>=1.0.0
langchain-mcp-adapters>=0.1.0
langgraph>=0.1.0
langchain-mistralai>=0.1.0
mcp-server-fetch>=0.1.0
